// src/global/Rubrique.d.ts
export interface Rubrique {
    
    designation: "string",
    enseignant: {
    adresse: "string",
    codePostal: "string",
      emailPerso: "string",
      emailUbo: "string",
      mobile: "string",
      noEnseignant: 0,
      nom: "string",
      pays: "string",
      prenom: "string",
      sexe: "string",
      telephone: "string",
      type: "string",
      ville: "string"
    },
    idRubrique: number,
    ordre: 0,
    type: "string"
  }